ff_ppc

+++ tasks/stl/laganina_e_component_labeling/src/ops_stl.cpp
@@ -70,7 +70,7 @@ bool laganina_e_component_labeling_stl::TestTaskSTL::PostProcessingImpl() {
 
 void laganina_e_component_labeling_stl::TestTaskSTL::InitializeParents(std::vector<int>& parent) {
   const int size = m_ * n_;
-  int num_threads = std::thread::hardware_concurrency();
+  int num_threads = static_cast<int>(std::thread::hardware_concurrency());
   // if (num_threads >= 6) {
   //   num_threads = 5;
   // }
@@ -95,7 +95,7 @@ void laganina_e_component_labeling_stl::TestTaskSTL::InitializeParents(std::vect
 
 void laganina_e_component_labeling_stl::TestTaskSTL::ProcessSweep(bool reverse, std::vector<int>& parent,
                                                                   bool& changed) const {
-  int num_threads = std::thread::hardware_concurrency();
+  int num_threads = static_cast<int>(std::thread::hardware_concurrency());
   // if (num_threads >= 6) {
   //   num_threads = 5;
   // }
@@ -190,7 +190,7 @@ bool laganina_e_component_labeling_stl::TestTaskSTL::UnionNodes(int a, int b, st
 
 void laganina_e_component_labeling_stl::TestTaskSTL::FinalizeRoots(std::vector<int>& parent) const {
   const int size = m_ * n_;
-  int num_threads = std::thread::hardware_concurrency();
+  int num_threads = static_cast<int>(std::thread::hardware_concurrency());
   // if (num_threads >= 6) {
   //   num_threads = 5;
   // }
@@ -217,7 +217,7 @@ void laganina_e_component_labeling_stl::TestTaskSTL::FinalizeRoots(std::vector<i
 
 void laganina_e_component_labeling_stl::TestTaskSTL::AssignLabels(std::vector<int>& parent) {
   const int size = m_ * n_;
-  int num_threads = std::thread::hardware_concurrency();
+  int num_threads = static_cast<int>(std::thread::hardware_concurrency());
   // if (num_threads >= 6) {
   //   num_threads = 5;
   // }
